+++ core/src/main/groovy/io/gatehill/imposter/exception/ResponseException.java
@@ -4,6 +4,10 @@ package io.gatehill.imposter.exception;
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 public class ResponseException extends RuntimeException {
+    public ResponseException(String message) {
+        super(message);
+    }
+
     public ResponseException(Exception cause) {
         super(cause);
     }