+++ plugin/openapi/src/main/java/io/gatehill/imposter/plugin/openapi/service/ExampleService.java
@@ -1,11 +1,11 @@
 package io.gatehill.imposter.plugin.openapi.service;
 
+import io.gatehill.imposter.ImposterConfig;
 import io.gatehill.imposter.plugin.openapi.config.OpenApiPluginConfig;
 import io.gatehill.imposter.script.ResponseBehaviour;
+import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.responses.ApiResponse;
 import io.vertx.ext.web.RoutingContext;
-
-import java.util.function.BiConsumer;
 
 /**
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
@@ -14,15 +14,17 @@ public interface ExampleService {
     /**
      * Attempt to respond with an example from the API specification.
      *
+     * @param imposterConfig    the Imposter engine configuration
      * @param config            the plugin configuration
      * @param routingContext    the Vert.x routing context
      * @param responseBehaviour the response behaviour
      * @param mockResponse      the specification response
-     * @param fallback          callback to invoke if no example was served
+     * @param spec              the OpenAPI specification
+     * @return {@code true} if an example was served, otherwise {@code false}
      */
-    void serveExample(OpenApiPluginConfig config,
-                      RoutingContext routingContext,
-                      ResponseBehaviour responseBehaviour,
-                      ApiResponse mockResponse,
-                      BiConsumer<RoutingContext, ResponseBehaviour> fallback);
+    boolean serveExample(ImposterConfig imposterConfig, OpenApiPluginConfig config,
+                         RoutingContext routingContext,
+                         ResponseBehaviour responseBehaviour,
+                         ApiResponse mockResponse,
+                         OpenAPI spec);
 }